Nilmani-backend/app/requirements.txt
# Core Dependencies for Interview Training Backend
requests
pydantic
pydantic-settings
python-dotenv

# FastAPI & Web Server
fastapi
uvicorn[standard]
python-multipart

# RAG & Vector Search
faiss-cpu
numpy
# PDF Processing
PyMuPDF

langchain
langchain-google-genai

# Optional: If you want to use official Ollama Python client (currently not used)
# ollama





# # Requirements for model_training/emotion_detection.ipynb
# # Minimal pinned versions to ensure evaluation_strategy/save_strategy and other APIs are available.
# # These are recommendations — if you need a GPU-specific wheel for torch, use the appropriate index as in the notebook.
# transformers>=4.11.0,<5.0.0
# datasets>=2.0.0
# huggingface-hub>=0.10.0
# accelerate>=0.20.0
# torch>=1.13.0
# torchaudio>=0.13.0
# librosa>=0.10.0
# soundfile>=0.12.1
# scikit-learn>=1.1.0
# numpy>=1.24
# seaborn>=0.12.0
# matplotlib>=3.5.0
# jiwer>=2.5.0
# noisereduce>=2.0.0
# packaging>=23.0
# tqdm>=4.65.0
# ipywidgets>=8.0.0
# torchvision>=0.14.0

# # Optional but helpful:
# # jupyter and ipywidgets are useful for notebook UI; install them if not present.
# jupyterlab>=3.5.0

# # Knowledge Base QA Requirements

# # FastAPI and server
# fastapi>=0.104.0
# uvicorn>=0.24.0
# python-multipart>=0.0.6

# # PDF processing
# PyMuPDF>=1.23.0
# PyPDF2>=3.0.0

# # Embeddings and vector search
# sentence-transformers>=2.2.2
# faiss-cpu>=1.7.4

# # LLM integration
# openai>=1.3.0

# # Data handling
# numpy>=1.24.0
# pydantic>=2.0.0
# python-dotenv>=1.0.0

# # Optional: For better sentence tokenization
# # nltk>=3.8.0
